+++ src/Instantia/CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
+## [0.2.0] - 2019-07-09
+### Added
+- `FromDateTimeOffset()` static method.
+- `GetCurrent()` static method as an alternative to `UtcNow` static property.
+### Removed
+- `Today` static property.
+
 ## [0.1.0] - 2019-06-30
 ### Added
 - `UtcDateTime` type.
 
+[0.2.0]: https://github.com/qbit86/instantia/compare/instantia-0.1.0...instantia-0.2.0
 [0.1.0]: https://github.com/qbit86/instantia/releases/tag/instantia-0.1.0